+++ libunhv3/bondchunk/fileinfo.cpp
@@ -1,3 +1,4 @@
+#include "unhv3.hpp"
 #include "bondchunkattr.hpp"
 #include "fileinfo.hpp"